feat(model-validation): address coderabit pr comments
